+++ src/map/battle.c
@@ -4483,7 +4483,7 @@ static struct Damage battle_calc_misc_attack(struct block_list *src, struct bloc
 			int64 matk=0, atk;
 			short tdef = status->get_total_def(target);
 			short tmdef =  status->get_total_mdef(target);
-			int targetVit = min(120, status_get_vit(target));
+			int targetVit = min(120, (int)status_get_vit(target));
 			short totaldef = (tmdef + tdef - ((uint64)(tmdef + tdef) >> 32)) >> 1; // FIXME: What's the >> 32 supposed to do here? tmdef and tdef are both 16-bit...
 
 			matk = battle->calc_magic_attack(src, target, skill_id, skill_lv, mflag).damage;
@@ -4503,7 +4503,7 @@ static struct Damage battle_calc_misc_attack(struct block_list *src, struct bloc
 
 				if( (vitfactor=(status_get_vit(target)-120.0f)) > 0)
 					vitfactor = (vitfactor * (matk + atk) / 10) / status_get_vit(target);
-				ftemp = max(0, vitfactor) + (targetVit * (matk + atk)) / 10;
+				ftemp = max(0.0f, vitfactor) + (targetVit * (matk + atk)) / 10;
 				md.damage = (int64)(ftemp * 70 * skill_lv / 100);
 				if (target->type == BL_PC)
 					md.damage >>= 1;
@@ -4636,7 +4636,7 @@ static struct Damage battle_calc_misc_attack(struct block_list *src, struct bloc
 		// (Base ATK + Weapon ATK) * Ratio
 		md.damage = (sstatus->batk + sstatus->rhw.atk) * (200 + 100 * skill_lv) / 100;
 		// Additional Damage
-		md.damage += sstatus->max_hp / (6 - min(5, skill_lv)) + status_get_max_sp(src) * (2 * skill_lv);
+		md.damage += sstatus->max_hp / (6 - min(5, (int)skill_lv)) + status_get_max_sp(src) * (2 * skill_lv);
 		break;
 	case SP_SOULEXPLOSION:
 		md.damage = tstatus->hp * (20 + 10 * skill_lv) / 100;
@@ -5141,7 +5141,7 @@ static struct Damage battle_calc_weapon_attack(struct block_list *src, struct bl
 						}
 				}
 				if ( wd.div_ > 1 ) {
-					wd.div_ = min(wd.div_, sd->status.inventory[i].amount);
+					wd.div_ = min(wd.div_, (int)sd->status.inventory[i].amount);
 					sc->data[SC_FEARBREEZE]->val4 = wd.div_ - 1;
 					wd.type = BDT_MULTIHIT;
 				}
@@ -6426,9 +6426,9 @@ static void battle_reflect_damage(struct block_list *target, struct block_list *
 	sc = status->get_sc(target);
 
 #ifdef RENEWAL
-#define NORMALIZE_RDAMAGE(d) ( trdamage += rdamage = max(1, min(max_reflect_damage, (d))) )
+#define NORMALIZE_RDAMAGE(d) ( trdamage += rdamage = max(1i64, min(max_reflect_damage, (d))) )
 #else
-#define NORMALIZE_RDAMAGE(d) ( trdamage += rdamage = max(1, (d)) )
+#define NORMALIZE_RDAMAGE(d) ( trdamage += rdamage = max(1i64, (d)) )
 #endif
 
 	if( sc && !sc->count )
